+++ lib/td_dd_web/controllers/data_structure_version_controller.ex
@@ -41,7 +41,8 @@ defmodule TdDdWeb.DataStructureVersionController do
     :versions,
     :with_confidential,
     :grant,
-    :grants
+    :grants,
+    :implementations
   ]
 
   swagger_path :show do