+++ test/td_dd_web/controllers/data_structure_version_controller_test.exs
@@ -669,6 +669,34 @@ defmodule TdDdWeb.DataStructureVersionControllerTest do
     end
   end
 
+  describe "GET /api/data_structures/:id/versions/:version implementations" do
+    setup :create_structure_with_implementation
+
+    @tag authentication: [role: "admin"]
+    test "rendes related implementations", %{
+      conn: conn,
+      data_structure: %{id: id},
+      implementation: %{implementation_key: implementation_key},
+      implementation_structure: %{id: implementation_structure_id}
+    } do
+      assert %{"data" => data} =
+               conn
+               |> get(
+                 Routes.data_structure_data_structure_version_path(conn, :show, id, "latest")
+               )
+               |> json_response(:ok)
+
+      assert %{
+               "implementations" => [
+                 %{
+                   "id" => ^implementation_structure_id,
+                   "implementation" => %{"implementation_key" => ^implementation_key}
+                 }
+               ]
+             } = data
+    end
+  end
+
   describe "bulk_update" do
     @tag authentication: [role: "admin"]
     test "bulk update of data structures", %{conn: conn} do
@@ -818,6 +846,24 @@ defmodule TdDdWeb.DataStructureVersionControllerTest do
       parent_version: parent_version,
       structure_version: structure_version,
       structure: structure
+    ]
+  end
+
+  defp create_structure_with_implementation(_) do
+    %{data_structure: data_structure} = insert(:data_structure_version)
+
+    implementation = insert(:implementation)
+
+    implementation_structure =
+      insert(:implementation_structure,
+        implementation: implementation,
+        data_structure: data_structure
+      )
+
+    [
+      implementation: implementation,
+      data_structure: data_structure,
+      implementation_structure: implementation_structure
     ]
   end