+++ test/td_dd_web/schema/data_structures_query_test.exs
@@ -1,0 +1,52 @@
+defmodule TdDdWeb.Schema.DataStructuresQueryTest do
+  use TdDdWeb.ConnCase
+
+  @moduletag sandbox: :shared
+
+  @query """
+  query LineageStructures($since: DateTime) {
+    dataStructures(since: $since, lineage: true) {
+      id
+      externalId
+      units {
+        id
+        name
+      }
+    }
+  }
+  """
+  @variables %{"since" => "2020-01-01T00:00:00Z"}
+
+  describe "dataStructures query" do
+    @tag authentication: [role: "user"]
+    test "returns forbidden when queried by user role", %{conn: conn} do
+      assert %{"data" => data, "errors" => errors} =
+               conn
+               |> post("/api/v2", %{"query" => @query, "variables" => %{}})
+               |> json_response(:ok)
+
+      assert data == %{"dataStructures" => nil}
+      assert [%{"message" => "forbidden"}] = errors
+    end
+
+    @tag authentication: [role: "service"]
+    test "returns data when queried by service role", %{conn: conn} do
+      %{data_structure: %{id: structure_id, external_id: external_id}} =
+        insert(:data_structure_version)
+
+      %{units: [%{name: unit_name}]} =
+        insert(:node, structure_id: structure_id, units: [build(:unit)])
+
+      assert %{"data" => data} =
+               response =
+               conn
+               |> post("/api/v2", %{"query" => @query, "variables" => @variables})
+               |> json_response(:ok)
+
+      assert response["errors"] == nil
+      assert %{"dataStructures" => data_structures} = data
+      assert [%{"id" => _, "externalId" => ^external_id, "units" => units}] = data_structures
+      assert [%{"id" => _, "name" => ^unit_name}] = units
+    end
+  end
+end