+++ CHANGELOG.md
@@ -5,11 +5,33 @@
 ### Added
 
 - [TD-4045] create_execution_group action
+- [TD-3186] Support for links between Implementations and Structures
 
-# Changed
+## [4.42.0] 2022-04-25
+
+### Added
+
+- [TD-4098] GraphQL query for data structures
+- [TD-4683] GraphQL query for data structure relations
+- [TD-2026] Add metadata to linage graph
+
+### Changed
 
 - [TD-4536] Support rule implementation with multiple populations
 
+### Fixed
+
+- [TD-4721] User with permissions can list sources
+- [TD-4783] Lineage Graph is not generated
+
+# Add
+
+- [TD-4695] Structure bulk update was not reindexing structures with updated
+  notes
+- [TD-3128] Profile `null_count`, `total_count` and `unique_count` fields were
+  limited to PostgreSQL `integer` values (maximum 2,147,483,647). They have now
+  been changed to `bigint` (maximum 9,223,372,036,854,775,807)
+- [TD-4538] Segments definitions in the implementations, and support for segment results
 ## [4.41.1] 2022-04-04
 
 # Added
@@ -22,6 +44,8 @@
 - [TD-4643] Raw implementation creation permission
 
 ## [4.41.0] 2022-03-28
+
+### Fixed
 
 - [TD-4535] Structures domains CSV upload bulk update